Match media types less strictly. Closes #1.
Media types are now matched at the level specified by the client. That
is, a response type may be negotiated that is MORE
specific (containing more type parameters) than requested by the
client in the Accept header.  This also applies to
content_types_accepted, meaning that an incoming entity may be
accepted by a less-specific processing method.

This will have less-surprising behavior when a client requests an
unparameterized type but the resource only provides parameterized
ones, as well as supporting patterns like versioned APIs via the type
parameters.