Test mongral adapter in integration
This actually caught an issue with the way response headers are handled.
Sometimes response headers can be array-valued, which was broken.

I also split the mongrel and reel adapters #run method into two parts,
in order to support graceful shutdown.

This fixes an annoying wart where ports are left open across tests, but
it also adds unnecessary complexity to the adapters.