Fix fiber response bodies in JRuby 1.7.4
The fiber response body adapter test failed for Reel, WEBrick, and Rack
(WEBrick backend) adapters.

I don't know why, but replacing the 3rd Fiber.yield with a simple return
value makes the fiber response body adapter tests pass.

The closest I got to finding the root cause was an exception handled in
WEBrick, which is passed to the server's logger's `error` method. The
exception is a java.lang.NullPointerException and the backtrace doesn't
include any lines from webmachine or the standard library.