add RackMapped adapter that correctly handles routing when hosted from a rack mapped URI
`Webmachine::Adapters::RackMapped` is implemented as a subclass of
`Webmachine::Adapters::Rack`, and simply overrides the way the
`Webmachine::Request` is built to explicitly set the `routing_tokens`
from the rack request `PATH_INFO` string - rather than from the full
uri path string as the base rack adapter does.

Rack provides the path that the app should use in the `PATH_INFO` env
var. This is the correct value to use for routing etc to work with
mapped apps.

Note, this was done as a separate class as suggested by @bethesque to
preserve backward compatibility. If any webmachine apps are running as
mapped rack apps already, they may have gotten around not being able to
host the app on a mapped path by specifying the full URI path in the
routing config - which would no longer work if this change was made on
the base rack adapter.