Revert back to before(:all) for adapter_lint
Using let-style RSpec and lazily creating our servers before every
test is good from a test purity point of view but increases greatly the
risk of port collisions when we attempt to find a free port to run
our adapter in adapter_lint (bearing in mind that Socket#close only
releases the Ruby resources, and the socket persists in a TIME_WAIT
state for an indeterminate period of time at OS level).

Revert to before(:all) and create only one server that will be used
for the duration of all tests in adapter_lint. We still need to sleep
 to ensure our 'free' ports are available, but we reduce the risk of
 servers being unavailable. This also has the happy side effect of
 making the tests somewhat faster.