chore: fix the build for ruby 3.0.x
* Add webrick as a dev dependency

webrick is no longer a bundled dependency for Ruby 3 and up.

* Fix translation errors in Ruby 3.x

Ruby 3 is stricter about hashes being used interchangeably with
keyword arguments, and will give errors like:

ArgumentError: wrong number of arguments (given 2, expected 0..1)

Explicitly map options hash to positional argument with ** to fix

* Replace Proc.new usage for Ruby 3

Ruby 3 no longer supports Proc.new without a block to mean "the block
that was passed to this method". Use an explicit &block instead
This commit does not fix as-notifications, on this line:

https://github.com/bernd/as-notifications/blob/v1.0.0/lib/as/notifications/fanout.rb#L18

* Update as-notifications to 1.0.2

Since 1.0.2 is now compatible with Ruby 3.0, we have one spec
we no longer have to use the Proc.new { } form with.

Thanks Bernd!