Add the standardrb linter
standardrb provides a relaxed set of linting rules that
provide the benefits of a linter without the headaches
that can come from the slightly excessive rubocop defaults.

New rules we'd like to adopt can be enabled in
`.rubocop.yml`, and similarly rules we don't want
to adopt can be disabled in `.rubocop.yml`.